Added support for using different types of resolvers.
